Move plans commands to plans subgroup
This commands make more sense in their own group, as other plans related
commands might be implemented.
